Add base Controller and Grade, Course/Exercise controllers and services; Add db config files;

+++ SofiaTeachersOnline/SofiaTeachersOnline.Database/Models/CourseProgress.cs
@@ -5,7 +5,8 @@ namespace SofiaTeachersOnline.Database.Models
 {
     public class CourseProgress : Entity
     {
-        public double Progress { get; set; }
+        // TODO: EntityController not working with this entity because it's not IModifiable
+        public float Progress { get; set; }
 
         public Guid StudentId { get; set; }
         public Student Student { get; set; }
